src/backend/fatura_api/requirements.txt
flask
pathlib
python-dotenv
PyJWT
Flask-SQLAlchemy
Flask-Migrate
flask-bcrypt
flask_jwt_extended
pymysql
requests
pytest
